third_party/WebKit/LayoutTests/transitions/unprefixed-perspective-origin.html

<!DOCTYPE html>
<div class="target">all 10ms</div><br>
<div class="target">perspective-origin 10ms</div><br>
<div class="target">-webkit-perspective-origin 10ms</div><br>
<div class="target">-webkit-perspective-origin-x 10ms</div><br>
<div class="target">-webkit-perspective-origin-y 10ms</div><br>
<script>
'use strict';
var remaining = 0;
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.waitUntilDone();
}

document.addEventListener('transitionend', function(e) {
  var result = document.createElement('div');
  result.textContent = 'fired transitionend with propertyName: ' + e.propertyName;
  e.target.appendChild(result);
  --remaining;
  if (remaining === 0 && window.testRunner) {
    testRunner.notifyDone();
  }
});

onload = function() {
  var targets = document.querySelectorAll('.target');
  for (var i = 0; i < targets.length; i++) {
    targets[i].offsetTop;
    targets[i].style.transition = targets[i].textContent;
    targets[i].style.webkitPerspectiveOriginX = '10px';
    targets[i].style.webkitPerspectiveOriginY = '10px';
    targets[i].style.perspectiveOrigin = '10px 10px';
    ++remaining;
  }
};
</script>
