third_party/blink/web_tests/transitions/unprefixed-perspective-origin-expected.txt

all 10ms
fired transitionend with propertyName: perspective-origin


perspective-origin 10ms
fired transitionend with propertyName: perspective-origin


-webkit-perspective-origin 10ms
fired transitionend with propertyName: perspective-origin


-webkit-perspective-origin-x 10ms
fired transitionend with propertyName: -webkit-perspective-origin-x


-webkit-perspective-origin-y 10ms
fired transitionend with propertyName: -webkit-perspective-origin-y
